Add files via upload
I found some datasets including the texts and their emotions (such as joy, angry, etc) and did some preprocessing. I used some NLP models (Bert based models) to find the emotion for each text that has highest entailment probability regarding relevance to the text as predicted emotion. I also used different description of emotions which are prompts or the average of synonyms of emotions as the representation of emotion. I also fine-tune models on the train sets and did prediction on test sets.